Moved authentication to server-side
