+++ index.js
@@ -1,9 +1,12 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 const db = require('./db');
+const { generateRandomString, pkceChallengeFromVerifier } = require('./src/utils/pkce_helper');
 
 const app = express();
 const port = 5000;
+let state = '';
+let codeVerifier = '';
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
@@ -62,4 +65,25 @@ app.get('/callback', (req, res) => {
     res.end();
     return status;
   }
+});
+
+app.get('/auth/login', async (req, res) => {
+  const config = {
+    client_id: '0oa2w7lue0U6fnn3N5d7',
+    redirect_uri: 'http://localhost:3000/',
+    authorization_endpoint: 'https://dev-88956181.okta.com/oauth2/default/v1/authorize',
+    token_endpoint: 'https://dev-88956181.okta.com/oauth2/default/v1/token',
+    request_scopes: 'openid',
+  };
+
+  // Create and store a random "state" value
+  state = generateRandomString();
+  // Create and store a new PKCE code_verifier (the plaintext random secret)
+  codeVerifier = generateRandomString();
+
+  const codeChallenge = await pkceChallengeFromVerifier(codeVerifier);
+
+  const url = `${config.authorization_endpoint}?response_type=code&client_id=${encodeURIComponent(config.client_id)}&state=${encodeURIComponent(state)}&scope=${encodeURIComponent(config.request_scopes)}&redirect_uri=${encodeURIComponent(config.redirect_uri)}&code_challenge=${encodeURIComponent(codeChallenge)}&code_challenge_method=S256`;
+
+  res.redirect(url);
 });
