+++ src/utils/pkce_helper.js
@@ -2,20 +2,18 @@
 - Securely generating a random string
 - Generating the SHA256 hash of that string
 */
+const crypto = require('crypto');
 
 // Generate a secure random string using the browser crypto functions
-export const generateRandomString = () => {
-  const array = new Uint32Array(28);
-  window.crypto.getRandomValues(array);
-  return Array.from(array, (dec) => (`0${dec.toString(16)}`).substr(-2)).join('');
-};
+const generateRandomString = () => crypto.randomBytes(20).toString('hex');
 
 // Calculate the SHA256 hash of the input text
 // Returns a promise that resolves to an ArrayBuffer.
-const sha256 = (input) => {
-  const encoder = new TextEncoder();
-  const data = encoder.encode(input);
-  return window.crypto.subtle.digest('SHA-256', data);
+const hash = (input) => {
+  console.log(input);
+  const hashedInput = crypto.createHash(input);
+  console.log(hashedInput);
+  return hashedInput;
 };
 
 // Base64-urlencodes the input string
@@ -25,7 +23,9 @@ const base64urlencode = (str) => (
 );
 
 // Return the base64-urlended sha256 hash for the PKCE Challenge
-export const pkceChallengeFromVerifier = async (v) => {
-  const hashed = await sha256(v);
+const pkceChallengeFromVerifier = async (v) => {
+  const hashed = await hash(v);
   return base64urlencode(hashed);
 };
+
+module.exports = { generateRandomString, pkceChallengeFromVerifier };
